Add the RecordWildcards extension, and remove a few
That long list of extensions is there to minimize the number of
MonadLoc errors the user will get to see.
If an extension is missing, let GHC be the one who reports it.

That said, I want to include in the list only those extensions which
I know for experience that don't mess up any syntax, at least
in practice. That's why e.g. TemplateHaskell goes out